requirements.txt
Flask==2.3.2
PyYAML==6.0.1
gunicorn==21.2.0
requests==2.31.0
